Added a header; Cleaned up some jsdocs; Added some new eslint rules
