--- apps/topic-inspector/src/app/_utils/to.ts
@@ -1,7 +1,0 @@
-export async function to<T = unknown, K = Error>(promise: () => Promise<T>) {
-    try {
-        return [await promise, null];
-    } catch (e) {
-        return [null, e as K];
-    }
-}
