+++ libs/utils/tsconfig.spec.json
@@ -1,0 +1,9 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+        "outDir": "../../dist/out-tsc",
+        "module": "commonjs",
+        "types": ["jest", "node"]
+    },
+    "include": ["**/*.test.ts", "**/*.spec.ts", "**/*.d.ts"]
+}
